chore: restructure project folders
