+++ src/shared/components/atoms/NextImage/INextImage.d.ts
@@ -1,0 +1,13 @@
+import type { ImageProps } from 'next/image';
+
+export type INextImageProps = {
+  useSkeleton?: boolean;
+  imgClassName?: string;
+  blurClassName?: string;
+  alt: string;
+  width: string | number;
+} & (
+  | { width: string | number; height: string | number }
+  | { layout: 'fill'; width?: string | number; height?: string | number }
+) &
+  ImageProps;
