+++ src/shared/components/atoms/index.ts
@@ -1,0 +1,2 @@
+export { default as NextImage } from './NextImage';
+export { default as Skeleton } from './Skeleton';
