+++ src/shared/components/molecules/index.ts
@@ -1,0 +1,1 @@
+export { default as Seo } from './Seo';
