+++ src/shared/lib/openGraph/IOpenGraph.d.ts
@@ -1,0 +1,6 @@
+export interface IOpenGraphType {
+  siteName: string;
+  description: string;
+  templateTitle?: string;
+  logo?: string;
+}
