+++ src/shared/lib/openGraph/index.ts
@@ -1,11 +1,8 @@
-type OpenGraphType = {
-  siteName: string;
-  description: string;
-  templateTitle?: string;
-  logo?: string;
-};
 // !STARTERCONF This OG is generated from https://github.com/theodorusclarence/og
 // Please clone them and self-host if your site is going to be visited by many people.
+
+import { IOpenGraphType } from './IOpenGraph';
+
 // Then change the url and the default logo.
 export function openGraph({
   siteName,
@@ -13,7 +10,7 @@ export function openGraph({
   description,
   // !STARTERCONF Or, you can use my server with your own logo.
   logo = 'https://og.thcl.dev/images/logo.jpg',
-}: OpenGraphType): string {
+}: IOpenGraphType): string {
   const ogLogo = encodeURIComponent(logo);
   const ogSiteName = encodeURIComponent(siteName.trim());
   const ogTemplateTitle = templateTitle
